Clear image_urls_ in FaviconHandler::FetchFavicon
So, it seems that FetchFavicon() is used to explicitly set the url, e.g. for a chrome:// page. If iamge_urls_ is not cleared, the fetch will not succeed. Clearing image_urls_ is safe here for the same reason it was safe in OnUpdateFaviconURL() - we check against the front of image_urls_ on any download so by clearing it any pending downloads will be safely and correctly ignored.

BUG=119601
TEST=See issue.


Review URL: http://codereview.chromium.org/9808091

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@128689 0039d316-1c4b-4281-b951-d872f2087c98