Fix range of gamepad axis values on Windows
Previously was being normalized via division by 32767, but the range is documented to be -32768..32767 so we need to bias to not exceed -1..1 and get the full range.

Also inspected the Linux code expecting a similar bug, but the range there is documented to be -32767..32767 so simple division suffices.

R=rsleevi
BUG=119791


Review URL: http://codereview.chromium.org/9838078

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@128773 0039d316-1c4b-4281-b951-d872f2087c98