Make json_schema_compiler resolve type names via resolution rules.
Instead of sharing a single namespace internally for types, makes
json_schema_compiler treat types as their fully-qualified selfs. This allows
two different namespaces to have types with the same name, and to be able to
reference the others' types by a fully-qualified name. This change also causes
namespaces to not be added twice to their CppTypeGenerators.

BUG=119954
TEST=none


Review URL: http://codereview.chromium.org/9834088

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@129251 0039d316-1c4b-4281-b951-d872f2087c98