aura: Change shared context to be offscreen for arm
ARM drivers don't like to have 2 EGLSurface on the same Window. So we need to
make the UI shared context offscren.
This also requires the image transport surface to be an "offscreen" GLSurface (a
pbuffer), however the GLES2Decoder needs to be "onscreen" (i.e. it calls
SwapBuffers on the surface instead managing the offscreen buffers itself).
So this decouples the 2 notions of "offscreen".

BUG=chrome-os-partner:8284
TEST=gpu_unittests. Aura chrome --ui-use-gpu-process with poster circle and webgl on tegra, lumpy, desktop GL. Desktop chrome with poster circle and webgl.


Review URL: http://codereview.chromium.org/10007034

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@131177 0039d316-1c4b-4281-b951-d872f2087c98