These targets were causing files to have the same output target, which ninja (reasonably) does not like.
dlldata.c is a generated file name that a bunch of idl build rules output but most do not actually compile it AFAICT. This was causing some targets in isimpledom (that contain multiple idl files) to output to the same name (dlldata.c), so uniquize it based on InputName. For the one target that seems to actually build dlldata.c, override back to the original name to keep things the same (iaccessible2).


Review URL: http://codereview.chromium.org/10008061

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@131232 0039d316-1c4b-4281-b951-d872f2087c98