Create Linux platform app shortcuts to run in their own process.
This change also reorganises the web_app::CreateShortcut code a bit to
allow us to do this without losing any functionality on gtk. Specifically
the ability to create a shortcut on the file thread and get a return value
is still available for Linux, and goes through the web_app code path.

Some of the variable names and comments in web_app.h were also updated
to be clearer.

BUG=None.
TEST=Test that shortcut creation works on Linux, both for platform apps and
  normal hosted apps.


Review URL: http://codereview.chromium.org/9958006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@132151 0039d316-1c4b-4281-b951-d872f2087c98