Use Value objects instead of JSON strings in extension API response IPCs.
When we're sending the result of an extension API method call, we currently
convert a Value object into a JSON string in the browser process, and then call
JSON.parse in javascript on the renderer side. This CL changes that to pass the
Value object in the IPC, and uses V8ValueConverter to turn the Value into the
appropriate v8 native object, avoiding the need to parse JSON.

This is a first step towards supporting binary data for API request
parameters and responses.

BUG=122675
TEST=No new functionality ; existing unit and browser tests should still pass.

Review URL: https://chromiumcodereview.appspot.com/10041047

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@133133 0039d316-1c4b-4281-b951-d872f2087c98