Actually hook up the Pepper clear site data.
This fixes the message routing. The message is sent over the "broker" channel in this case which normally connects to the renderer. Switching this over to be the main plugin <-> browser channel is extra work since we'd need a way to plumb through the response. Since we already do this weird routing for the NPAPI plugin, doing the same thing is the simplest here.

I moved the code to the broker dispatcher which is the endpoint for the channel that's used.

This removes the timeout for sending the clear message. This made it challenging to debug since if you load the plugin in the debugger, the message never gets sent. This could legitimately happen if the computer is swapping and is super slow to load the plugin. It seems if you want to clear your data, we shouldn't "forget" to clear it in certain cases. So I just left this code out.
Review URL: https://chromiumcodereview.appspot.com/10155001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@133277 0039d316-1c4b-4281-b951-d872f2087c98