NaCl: Attach the Windows debug exception handler on demand
This uses the callback that I added on the NaCl side in an earlier
change.

We add a "success" parameter to various interfaces to indicate whether
the debug exception handler was attached successfully.

With this change, the convoluted logic for delaying the start message
goes away.  We remove the DebugContext class.  It is replaced by the
implementation in chrome/common/nacl_debug_exception_handler_win.cc.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2618
TEST=run_inbrowser_exception_test in nacl_integration


Review URL: http://codereview.chromium.org/10267013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@135105 0039d316-1c4b-4281-b951-d872f2087c98