Chrome Endure tests now monitor for garbage collection events in the v8 heap.
When garbage collection events occur, information about those events is
outputted in a format that can be graphed by the Chrome Endure graphs.

BUG=chromium-os:23967
TEST=Verified the Gmail compose/discard Chrome Endure test still runs
successfully on my local machine with these changes, and that the information
associated with garbage collections is outputted by the test when garbage
collections occur in the v8 heap.

Review URL: https://chromiumcodereview.appspot.com/10201004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@135168 0039d316-1c4b-4281-b951-d872f2087c98